+++ ginkgo/testrunner/test_runner.go
@@ -22,8 +22,10 @@ import (
 )
 
 type TestRunner struct {
-	Suite    testsuite.TestSuite
-	compiled bool
+	Suite testsuite.TestSuite
+
+	compiled              bool
+	compilationTargetPath string
 
 	numCPU         int
 	parallelStream bool
@@ -34,7 +36,7 @@ type TestRunner struct {
 }
 
 func New(suite testsuite.TestSuite, numCPU int, parallelStream bool, race bool, cover bool, tags string, additionalArgs []string) *TestRunner {
-	return &TestRunner{
+	runner := &TestRunner{
 		Suite:          suite,
 		numCPU:         numCPU,
 		parallelStream: parallelStream,
@@ -43,9 +45,23 @@ func New(suite testsuite.TestSuite, numCPU int, parallelStream bool, race bool, 
 		tags:           tags,
 		additionalArgs: additionalArgs,
 	}
+
+	if !suite.Precompiled {
+		dir, err := ioutil.TempDir("", "ginkgo")
+		if err != nil {
+			panic(fmt.Sprintf("coulnd't create temporary directory... might be time to rm -rf:\n%s", err.Error()))
+		}
+		runner.compilationTargetPath = filepath.Join(dir, suite.PackageName+".test")
+	}
+
+	return runner
 }
 
 func (t *TestRunner) Compile() error {
+	return t.CompileTo(t.compilationTargetPath)
+}
+
+func (t *TestRunner) CompileTo(path string) error {
 	if t.compiled {
 		return nil
 	}
@@ -54,9 +70,7 @@ func (t *TestRunner) Compile() error {
 		return nil
 	}
 
-	os.Remove(t.compiledArtifact())
-
-	args := []string{"test", "-c", "-i"}
+	args := []string{"test", "-c", "-i", "-o", path}
 	if t.race {
 		args = append(args, "-race")
 	}
@@ -78,10 +92,11 @@ func (t *TestRunner) Compile() error {
 		if len(output) > 0 {
 			return fmt.Errorf("Failed to compile %s:\n\n%s", t.Suite.PackageName, fixedOutput)
 		}
-		return fmt.Errorf("")
+		return fmt.Errorf("Failed to compile %s", t.Suite.PackageName)
 	}
 
 	t.compiled = true
+
 	return nil
 }
 
@@ -134,12 +149,7 @@ func (t *TestRunner) CleanUp() {
 	if t.Suite.Precompiled {
 		return
 	}
-	os.Remove(t.compiledArtifact())
-}
-
-func (t *TestRunner) compiledArtifact() string {
-	compiledArtifact, _ := filepath.Abs(filepath.Join(t.Suite.Path, fmt.Sprintf("%s.test", t.Suite.PackageName)))
-	return compiledArtifact
+	os.RemoveAll(filepath.Dir(t.compilationTargetPath))
 }
 
 func (t *TestRunner) runSerialGinkgoSuite() RunResult {
@@ -303,7 +313,12 @@ func (t *TestRunner) cmd(ginkgoArgs []string, stream io.Writer, node int) *exec.
 	args = append(args, ginkgoArgs...)
 	args = append(args, t.additionalArgs...)
 
-	cmd := exec.Command(t.compiledArtifact(), args...)
+	path := t.compilationTargetPath
+	if t.Suite.Precompiled {
+		path, _ = filepath.Abs(filepath.Join(t.Suite.Path, fmt.Sprintf("%s.test", t.Suite.PackageName)))
+	}
+
+	cmd := exec.Command(path, args...)
 
 	cmd.Dir = t.Suite.Path
 	cmd.Stderr = stream